Trait vortex_serde::layouts::LayoutReader

pub trait LayoutReader: Debug + Send {
    // Required method
    fn read_next(&mut self) -> VortexResult<Option<ReadResult>>;
}

Required Methods

fn read_next(&mut self) -> VortexResult<Option<ReadResult>>

Reads the data from the underlying layout

The layout can either return a batch data, i.e. an Array or ask for more layout messages to be read. When requesting messages to be read the caller should populate the messages in the cache and then call back into this function.

The layout is finished reading when it returns None